+++ mlfinlab/data_structures/base_bars.py
@@ -204,3 +204,118 @@ class BaseBars(ABC):
             raise ValueError('Unknown metric')
 
         return imbalance
+
+class BaseImbalanceBars(BaseBars):
+    def __init__(self, file_path, metric, batch_size, expected_imbalance_window, expected_num_ticks_init, analyse_thresholds):
+        BaseBars.__init__(self, file_path, metric, batch_size)
+
+        self.expected_imbalance_window = expected_imbalance_window
+        # Expected number of ticks extracted from prev bars
+        self.exp_num_ticks = self.exp_num_ticks_init
+        self.num_ticks_bar = []  # List of number of ticks from previous bars
+
+        self.cum_theta = 0
+        self.expected_imbalance = np.nan
+        self.imbalance_array = []
+
+        self.analyse_thresholds = analyse_thresholds
+        self.bars_thresholds = []  # Array of dicts: {'timestamp': value, 'cum_theta': value, 'exp_num_ticks': value, 'exp_imbalance': value}
+
+    def _reset_cache(self):
+        """
+        Implementation of abstract method _reset_cache for imbalance bars
+        """
+        self.open_price = None
+        self.high_price, self.low_price = -np.inf, np.inf
+        self.cum_ticks, self.cum_dollar_value, self.cum_volume, self.cum_theta = 0, 0, 0, 0
+
+    def _extract_bars(self, data):
+        """
+        For loop which compiles the various imbalance bars: dollar, volume, or tick.
+
+        :param data: (DataFrame) Contains 3 columns - date_time, price, and volume.
+        :return: (List) of bars built using the current batch.
+        """
+
+        # Iterate over rows
+        list_bars = []
+        for row in data.values:
+            # Set variables
+            date_time = row[0]
+            price = np.float(row[1])
+            volume = row[2]
+
+            self.cum_ticks += 1
+
+            if self.open_price is None:
+                self.open_price = price
+
+            self.cum_volume += volume
+
+            # Update high low prices
+            self.high_price, self.low_price = self._update_high_low(
+                self.high_price, self.low_price, price)
+
+            # Imbalance calculations
+            signed_tick = self._apply_tick_rule(price)
+            imbalance = self._get_imbalance(price, signed_tick, volume)
+            self.imbalance_array.append(imbalance)
+            self.cum_theta += imbalance
+
+            # Get expected imbalance for the first time, when num_ticks_init passed
+            if not list_bars and np.isnan(self.expected_imbalance):
+                self.expected_imbalance = self._get_expected_imbalance(
+                    self.expected_imbalance_window)
+
+            if self.analyse_thresholds is True:
+                self.bars_thresholds.append(
+                    {'timestamp': date_time, 'cum_theta': self.cum_theta, 'exp_num_ticks': self.exp_num_ticks,
+                     'exp_imbalance': np.abs(self.expected_imbalance)})
+
+            self.prev_price = price  # Update previous price used for tick rule calculations
+
+            # Check expression for possible bar generation
+            if np.abs(self.cum_theta) > self.exp_num_ticks * np.abs(self.expected_imbalance):
+                self._create_bars(date_time, price,
+                                  self.high_price, self.low_price, list_bars)
+
+                self.num_ticks_bar.append(self.cum_ticks)
+                # Expected number of ticks based on formed bars
+                self.exp_num_ticks = self._get_exp_num_ticks()
+                # Get expected imbalance
+                self.expected_imbalance = self._get_expected_imbalance(
+                    self.expected_imbalance_window)
+                # Reset counters
+                self._reset_cache()
+
+        return list_bars
+
+    def _get_expected_imbalance(self, window):
+        """
+        Calculate the expected imbalance: 2P[b_t=1]-1, using a EWMA, pg 29
+        :param window: EWMA window for calculation
+        :return: expected_imbalance: 2P[b_t=1]-1, approximated using a EWMA
+        """
+        if len(self.imbalance_array) < self.exp_num_ticks_init:
+            # Waiting for array to fill for ewma
+            ewma_window = np.nan
+        else:
+            # ewma window can be either the window specified in a function call
+            # or it is len of imbalance_array if window > len(imbalance_array)
+            ewma_window = int(min(len(self.imbalance_array), window))
+
+        if np.isnan(ewma_window):
+            # return nan, wait until len(self.imbalance_array) >= self.exp_num_ticks_init
+            expected_imbalance = np.nan
+        else:
+            expected_imbalance = ewma(
+                np.array(self.imbalance_array[-ewma_window:], dtype=float), window=ewma_window)[-1]
+
+        return expected_imbalance
+
+    @abstractmethod
+    def _get_exp_num_ticks(self):
+        """
+        """
+
+    def
